Cleanup / rearrange EntitySubset files
