+++ python_modules/dagster-graphql/dagster_graphql/schema/asset_condition_evaluations.py
@@ -3,10 +3,10 @@ import itertools
 from typing import Optional, Sequence
 
 import graphene
+from dagster._core.asset_graph_view.serializable_entity_subset import SerializableEntitySubset
 from dagster._core.definitions.declarative_automation.serialized_objects import (
     AutomationConditionEvaluation,
 )
-from dagster._core.definitions.entity_subset import SerializableEntitySubset
 from dagster._core.definitions.partition import PartitionsDefinition
 from dagster._core.scheduler.instigation import AutoMaterializeAssetEvaluationRecord
