+++ python_modules/dagster/dagster/_core/definitions/auto_materialize_rule_impls.py
@@ -15,6 +15,7 @@ from typing import (
 )
 
 from dagster._annotations import deprecated, experimental
+from dagster._core.asset_graph_view.serializable_entity_subset import SerializableEntitySubset
 from dagster._core.definitions.auto_materialize_rule import AutoMaterializeRule
 from dagster._core.definitions.auto_materialize_rule_evaluation import (
     AutoMaterializeDecisionType,
@@ -25,7 +26,6 @@ from dagster._core.definitions.base_asset_graph import sort_key_for_asset_partit
 from dagster._core.definitions.declarative_automation.legacy.valid_asset_subset import (
     ValidAssetSubset,
 )
-from dagster._core.definitions.entity_subset import SerializableEntitySubset
 from dagster._core.definitions.events import AssetKey, AssetKeyPartitionKey
 from dagster._core.definitions.freshness_based_auto_materialize import (
     freshness_evaluation_results_for_asset_key,
