+++ python_modules/dagster/dagster/_core/definitions/declarative_automation/operators/newly_true_operator.py
@@ -1,6 +1,7 @@
 from typing import Optional, Sequence
 
-from dagster._core.asset_graph_view.asset_graph_view import EntitySubset
+from dagster._core.asset_graph_view.entity_subset import EntitySubset
+from dagster._core.asset_graph_view.serializable_entity_subset import SerializableEntitySubset
 from dagster._core.definitions.asset_key import T_EntityKey
 from dagster._core.definitions.declarative_automation.automation_condition import (
     AutomationCondition,
@@ -8,7 +9,6 @@ from dagster._core.definitions.declarative_automation.automation_condition impor
     BuiltinAutomationCondition,
 )
 from dagster._core.definitions.declarative_automation.automation_context import AutomationContext
-from dagster._core.definitions.entity_subset import SerializableEntitySubset
 from dagster._serdes.serdes import whitelist_for_serdes
