+++ python_modules/dagster/dagster/_utils/test/schedule_storage.py
@@ -4,12 +4,12 @@ import time
 import pytest
 
 from dagster import StaticPartitionsDefinition
+from dagster._core.asset_graph_view.serializable_entity_subset import SerializableEntitySubset
 from dagster._core.definitions.declarative_automation.serialized_objects import (
     AssetSubsetWithMetadata,
     AutomationConditionEvaluation,
     AutomationConditionNodeSnapshot,
 )
-from dagster._core.definitions.entity_subset import SerializableEntitySubset
 from dagster._core.definitions.events import AssetKey
 from dagster._core.definitions.metadata import MetadataValue
 from dagster._core.remote_representation import (
