+++ python_modules/dagster/dagster_tests/asset_defs_tests/test_entity_subset.py
@@ -11,10 +11,10 @@ from dagster import (
     PartitionsDefinition,
     StaticPartitionsDefinition,
 )
+from dagster._core.asset_graph_view.serializable_entity_subset import SerializableEntitySubset
 from dagster._core.definitions.declarative_automation.legacy.valid_asset_subset import (
     ValidAssetSubset,
 )
-from dagster._core.definitions.entity_subset import SerializableEntitySubset
 from dagster._core.definitions.events import AssetKeyPartitionKey
 from dagster._core.definitions.partition import AllPartitionsSubset, DefaultPartitionsSubset
 from dagster._core.definitions.time_window_partitions import (
